Refactor Work templates for simplicity and clarity.
